requirements.txt
autopep8==1.5.4
click==7.1.2
dnspython==1.16.0
docformatter==1.3.1
fastapi==0.61.0
gunicorn==20.0.4
h11==0.9.0
linecache2==1.0.0
numpy==1.19.1
pandas==1.1.1
pycodestyle==2.6.0
pydantic==1.6.1
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
starlette==0.13.6
toml==0.10.1
traceback2==1.4.0
unittest2==1.1.0
untokenize==0.1.1
uvicorn==0.11.8
websockets==8.1
